added start_process overload stating startup_info to process.h
- some warning clean up
- added using alias' for ansi_process_startup_info_builder and wide_process_startup_info_builder
