+++ include/modern_win32/process.h
@@ -18,7 +18,7 @@
 #include <modern_win32/modern_win32_export.h>
 #include <modern_win32/null_handle.h>
 #include <modern_win32/process_enums.h>
-#include <modern_win32/shared_utilities.h>
+#include <modern_win32/process_startup_info.h>
 
 #include <chrono>
 #include <optional>
@@ -157,7 +157,7 @@ namespace modern_win32
     /// <summary>
     /// Opens an existing local process object.
     /// </summary>
-    /// <param name="process_id">The identifier of the local process to be opened.</param>
+    /// <param name="id">The identifier of the local process to be opened.</param>
     /// <param name="access_rights">
     /// The access to the process object. This access right is checked against the security
     /// descriptor for the process. This parameter can be one or more of the <see cref="process_access_rights"/>.
@@ -168,7 +168,42 @@ namespace modern_win32
     /// <returns></returns>
     /// <exception cref="std::invalid_argument">if <paramref name="process_id"/> is 0</exception>
     /// <exception cref="access_denied_exception">if insufficent access to open process</exception>
-    [[nodiscard]] MODERN_WIN32_EXPORT process open_process(process_id_type const& process_id, process_access_rights const access_rights, bool const inherit_handles = false);
+    [[nodiscard]] MODERN_WIN32_EXPORT process open_process(process_id_type const& id, process_access_rights const access_rights, bool const inherit_handles = false);
+
+    /// <summary>
+    /// Starts the process resource that is specified by the parameter containing process start information
+    /// (for example, the file name of the process to start) and associates the resource with a new Process component.
+    /// </summary>
+    /// <param name="startup_info">
+    /// The <see cref="ansi_process_start_info"/> that contains the information that is used to start the process,
+    /// including the file name and any command-line arguments.
+    /// </param>
+    /// <returns>
+    /// A new Process that is associated with the process resource, Note that a new process that's started
+    /// alongside already running instances of the same process will be independent from the others. In addition,
+    /// start may return a non-null process with its has_exited already set to true. In this case, the started
+    /// process may have activated an existing instance of itself and then exited.
+    /// </returns>
+    /// <exception cref="std::filesystem::filesystem_error">thrown if thte filename specified in <paramref name="startup_info"/> is not found.</exception>
+    /// <exception cref="windows_exception">if there is an error in the native CreateProcess call</exception>
+    [[nodiscard]] process start_process(ansi_process_startup_info const& startup_info);
+    /// <summary>
+    /// Starts the process resource that is specified by the parameter containing process start information
+    /// (for example, the file name of the process to start) and associates the resource with a new Process component.
+    /// </summary>
+    /// <param name="startup_info">
+    /// The <see cref="wide_process_start_info"/> that contains the information that is used to start the process,
+    /// including the file name and any command-line arguments.
+    /// </param>
+    /// <returns>
+    /// A new Process that is associated with the process resource, Note that a new process that's started
+    /// alongside already running instances of the same process will be independent from the others. In addition,
+    /// start may return a non-null process with its has_exited already set to true. In this case, the started
+    /// process may have activated an existing instance of itself and then exited.
+    /// </returns>
+    /// <exception cref="std::filesystem::filesystem_error">thrown if thte filename specified in <paramref name="startup_info"/> is not found.</exception>
+    /// <exception cref="windows_exception">if there is an error in the native CreateProcess call</exception>
+    [[nodiscard]] process start_process(wide_process_startup_info const& startup_info);
 
     /// <summary>
     /// Starts a process resource by specifying the name of an application and
