+++ src/modern_win32/process.cpp
@@ -33,7 +33,7 @@ process::process(process_id_type const& id)
         throw std::invalid_argument("id cannot be 0");
 }
 process::process(process_id_type const& id, process_access_rights const access_rights, bool const inherit_handles)
-    : process(m_id)
+    : process(id)
 {
     m_handle = impl::get_process_handle(m_id, access_rights, inherit_handles);
 }
@@ -220,9 +220,9 @@ void process::close() noexcept
 
 }
 
-process open_process(process_id_type const& process_id, process_access_rights const access_rights, bool const inherit_handles)
+process open_process(process_id_type const& id, process_access_rights const access_rights, bool const inherit_handles)
 {
-    process process{process_id};
+    process process{id, access_rights, inherit_handles};
 
     if (static_cast<bool>(process))
         return process;
@@ -238,109 +238,18 @@ process open_process(process_id_type const& process_id, process_access_rights co
     }
 }
 
-template <typename TCHAR>
-std::vector<TCHAR>&& build_environment_block(typename process_startup_info<TCHAR>::environment_variable_container const& maybe_environment_map)
+process start_process(ansi_process_startup_info const& startup_info)
 {
-    std::vector<TCHAR> environment_block;
-    if (!maybe_environment_map.has_value())
-        return std::move(environment_block);
-
-    auto& environment_map = maybe_environment_map.value();
-    auto const nul_char = TCHAR(0);
-    auto const equals_char = TCHAR('=');
-
-    using std::copy;
-    using std::back_inserter;
-
-    for (auto& pair : environment_map)
-    {
-        copy(begin(pair.first), end(pair.first), back_inserter(environment_block));
-        environment_block.push_back(equals_char);
-        copy(begin(pair.second), end(pair.second), back_inserter(environment_block));
-        environment_block.push_back(nul_char);
-    }
-    if (!environment_block.empty())
-        environment_block.push_back(nul_char);
-
-    return std::move(environment_block);
+    return impl::start_process<char>(startup_info);
 }
-
-template <typename TCHAR>
-[[nodiscard]] bool create_process_t(TCHAR* command_line, bool inherit_handles, 
-    std::vector<TCHAR>& environment_block, TCHAR const* directory, 
-    process_startup_info<TCHAR> const& startup_info, PROCESS_INFORMATION& process_information)
+process start_process(wide_process_startup_info const& startup_info)
 {
-    return false;
-}
-
-template <>
-[[nodiscard]] bool create_process_t(char* command_line, bool const inherit_handles, 
-    std::vector<char>& environment_block, char const* directory, 
-    process_startup_info<char> const& startup_info, PROCESS_INFORMATION& process_information)
-{
-    auto native_startup_info = startup_info.build_native_startup_info<STARTUPINFOA>();
-    return CreateProcessA(
-        nullptr,
-        command_line, 
-        nullptr, 
-        nullptr, 
-        inherit_handles ? TRUE : FALSE, 
-        to_underlying_type(startup_info.build_creation_options()), 
-        environment_block.empty() ? static_cast<void*>(nullptr) : static_cast<void*>(&environment_block[0]),
-        directory,
-        &native_startup_info, 
-        &process_information) == TRUE;
-}
-template <>
-[[nodiscard]] bool create_process_t(wchar_t* command_line, bool const inherit_handles, 
-    std::vector<wchar_t>& environment_block, wchar_t const* directory, 
-    process_startup_info<wchar_t> const& startup_info, PROCESS_INFORMATION& process_information)
-{
-    auto native_startup_info = startup_info.build_native_startup_info<STARTUPINFOW>();
-    return CreateProcessW(
-        nullptr,
-        command_line, 
-        nullptr, 
-        nullptr, 
-        inherit_handles ? TRUE : FALSE, 
-        to_underlying_type(startup_info.build_creation_options()), 
-        environment_block.empty() ? static_cast<void*>(nullptr) : static_cast<void*>(&environment_block[0]),
-        directory,
-        &native_startup_info, 
-        &process_information) == TRUE;
-}
-
-template <typename TCHAR>
-process start_process(process_startup_info<TCHAR> const& startup_info)
-{
-    std::filesystem::path const filename(startup_info.filename);
-    if (!exists(filename))
-        throw std::filesystem::filesystem_error("filename not found", std::error_code(static_cast<int>(windows_error::error_file_not_found), std::iostream_category()));
-
-    auto command_buffer = startup_info.build_command_buffer();
-
-    PROCESS_INFORMATION process_information{};
-
-    auto environment_block = build_environment_block<TCHAR>(startup_info.environment);
-
-    auto const result = create_process_t<TCHAR>(
-        command_buffer.get(), 
-        startup_info.inherit_handles,
-        environment_block,
-        startup_info.directory.empty() ? nullptr : startup_info.directory.c_str(), // working directory,
-        startup_info,
-        process_information);
-
-    if (!result)
-        throw windows_exception();
-
-    CloseHandle(process_information.hThread);
-    return process(process_information.dwProcessId, process_information.hProcess);
+    return impl::start_process<wchar_t>(startup_info);
 }
 
 process start_process(char const* filename, char const* arguments)
 {
-    auto const startup_info = process_startup_info_builder<char>()
+    auto const startup_info = ansi_process_startup_info_builder()
         .with_filename(filename)
         .with_arguments(arguments)
         .build();
@@ -349,11 +258,11 @@ process start_process(char const* filename, char const* arguments)
 
 process start_process(wchar_t const* filename, wchar_t const* arguments)
 {
-    auto const startup_info = process_startup_info_builder<wchar_t>()
+    auto const startup_info = wide_process_startup_info_builder()
         .with_filename(filename)
         .with_arguments(arguments)
         .build();
-    return start_process<wchar_t>(startup_info);
+    return start_process(startup_info);
 
 }
